refactor(ErrorMessage): ErrorMessage 클래스 패키지 변경 및 사용하지 않는 패키지 삭제
