+++ src/main/java/domain/validator/ErrorMessage.java
@@ -1,4 +1,4 @@
-package validation;
+package domain.validator;
 
 public enum ErrorMessage {
